Fix DORA metric Graphql type
Some DORA metrics can be float
so we need float data type

Changelog: fixed
EE: true

+++ ee/app/graphql/types/dora_metric_type.rb
@@ -7,7 +7,7 @@ module Types
 
     field :date, GraphQL::Types::String, null: true,
           description: 'Date of the data point.'
-    field :value, GraphQL::Types::Int, null: true,
+    field :value, GraphQL::Types::Float, null: true,
           description: 'Value of the data point.'
   end
   # rubocop: enable Graphql/AuthorizeTypes
